relocate Moods to components folder
